crypto: put android crypto footer on new crypto partition
This places the crypto footer used by Android's crypto implementation
into a new partition "crypto" created to for this purpose.
This footer needs to be in unformatted storage that is read/write.

The Android crypto footer contains details on the type of encryption, and an
encrypted copy of the master key to decrypt the filesystem. Android uses
this footer without a filesystem, directly accessing the block device.

Change-Id: I9d1839d032c70658313137842286d19df27eaa5c
Signed-off-by: Bryan Buckley <bryan.buckley@ti.com>